+++ sharding-core/sharding-core-route/src/main/java/org/apache/shardingsphere/core/route/router/sharding/validator/statement/impl/ShardingInsertValidator.java
@@ -15,14 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.route.router.sharding.validator.impl;
+package org.apache.shardingsphere.core.route.router.sharding.validator.statement.impl;
 
 import com.google.common.base.Optional;
 import org.apache.shardingsphere.core.exception.ShardingException;
 import org.apache.shardingsphere.core.parse.sql.segment.dml.column.ColumnSegment;
 import org.apache.shardingsphere.core.parse.sql.segment.dml.column.OnDuplicateKeyColumnsSegment;
 import org.apache.shardingsphere.core.parse.sql.statement.dml.InsertStatement;
-import org.apache.shardingsphere.core.route.router.sharding.validator.ShardingValidator;
+import org.apache.shardingsphere.core.route.router.sharding.validator.statement.ShardingValidator;
 import org.apache.shardingsphere.core.rule.ShardingRule;
 
 /**
