+++ sharding-core/sharding-core-route/src/main/java/org/apache/shardingsphere/core/route/router/sharding/validator/statement/impl/ShardingUpdateValidator.java
@@ -15,13 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.route.router.sharding.validator.impl;
+package org.apache.shardingsphere.core.route.router.sharding.validator.statement.impl;
 
 import org.apache.shardingsphere.core.exception.ShardingException;
 import org.apache.shardingsphere.core.optimize.api.segment.Tables;
 import org.apache.shardingsphere.core.parse.sql.segment.dml.assignment.AssignmentSegment;
 import org.apache.shardingsphere.core.parse.sql.statement.dml.UpdateStatement;
-import org.apache.shardingsphere.core.route.router.sharding.validator.ShardingValidator;
+import org.apache.shardingsphere.core.route.router.sharding.validator.statement.ShardingValidator;
 import org.apache.shardingsphere.core.rule.ShardingRule;
 
 /**
